refactor(MeetingList): 불필요한 로직 제거
#36
